+++ python/src/lib.rs
@@ -10,7 +10,7 @@ use dbn::{
         BidAskPair, ErrorMsg, ImbalanceMsg, InstrumentDefMsg, MboMsg, Mbp10Msg, Mbp1Msg, OhlcvMsg,
         RecordHeader, StatMsg, StatusMsg, SymbolMappingMsg, SystemMsg, TradeMsg,
     },
-    Metadata, FIXED_PRICE_SCALE, UNDEF_ORDER_SIZE, UNDEF_PRICE, UNDEF_STAT_QUANTITY,
+    Metadata, RType, FIXED_PRICE_SCALE, UNDEF_ORDER_SIZE, UNDEF_PRICE, UNDEF_STAT_QUANTITY,
     UNDEF_TIMESTAMP,
 };
 
@@ -53,6 +53,7 @@ fn databento_dbn(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
     checked_add_class::<Encoding>(m)?;
     checked_add_class::<Schema>(m)?;
     checked_add_class::<SType>(m)?;
+    checked_add_class::<RType>(m)?;
     // constants
     m.add("FIXED_PRICE_SCALE", FIXED_PRICE_SCALE)?;
     m.add("UNDEF_PRICE", UNDEF_PRICE)?;