+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 - Added new publisher values in preparation for IFEU.IMPACT and NDEX.IMPACT datasets
 - Added `MAX_RECORD_LEN` constant for the length of the largest record type
 - Exposed record flag constants in `databento_dbn` with `F_` prefix
+- Added export to Python for `RType`
 
 ### Breaking changes
 - The old `InstrumentDefMsg` is now `compat::InstrumentDefMsgV1`